+++ packages/cli/CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ## Unreleased
 
+#### 🚀 Updates
+
+- Added `hasher.batchSize` to control the number of files to be hashed per batch.
+
 #### 🐞 Fixes
 
 - Fixed an issue where non-input matching files were being passed to `git hash-object` during the